Add stack-based profile flows with slide transitions
