+++ mobile/app/(client)/(profile)/profileDetails.tsx
@@ -10,44 +10,20 @@ import {
   KeyboardAvoidingView,
   Platform,
 } from "react-native";
-import { Stack, router, useLocalSearchParams } from "expo-router";
+import { Stack, router } from "expo-router";
 import * as ImagePicker from "expo-image-picker";
 import { Ionicons } from "@expo/vector-icons";
 import { Colors } from "@src/theme/tokens";
 import { useAuth } from "@src/store/useAuth";
-import { useProfile, type RoleKey } from "@src/store/useProfile";
+import { useProfile } from "@src/store/useProfile";
 import { uploadAvatar, uploadBanner } from "@src/lib/api";
 import { useSafeAreaInsets } from "react-native-safe-area-context";
 
-const BACK_TO = "/(labourer)/profile";
 
-export default function LabourerProfileDetails() {
+export default function ClientProfileDetails() {
   const insets = useSafeAreaInsets();
   const { user, token } = useAuth();
-  const authUserId = user?.id ?? 0;
-  const params = useLocalSearchParams<{
-    userId?: string;
-    jobId?: string;
-    from?: string;
-    role?: string;
-  }>();
-  const viewUserId = params.userId
-    ? parseInt(Array.isArray(params.userId) ? params.userId[0] : params.userId, 10)
-    : authUserId;
-  const backJobId = params.jobId
-    ? Array.isArray(params.jobId) ? params.jobId[0] : params.jobId
-    : undefined;
-  const from = params.from
-    ? Array.isArray(params.from)
-      ? params.from[0]
-      : params.from
-    : undefined;
-  const viewRole = (params.role
-    ? Array.isArray(params.role)
-      ? params.role[0]
-      : params.role
-    : "manager") as RoleKey;
-  const isOwn = viewUserId === authUserId;
+  const userId = user?.id ?? 0;
 
   const profiles = useProfile((s) => s.profiles);
   const ensureProfile = useProfile((s) => s.ensureProfile);
@@ -61,33 +37,17 @@ export default function LabourerProfileDetails() {
   const removeQualification = useProfile((s) => s.removeQualification);
 
   useEffect(() => {
-    if (!profiles[viewUserId]) {
-      if (isOwn && user) {
-        ensureProfile(
-          viewUserId,
-          user.username ?? "You",
-          (user.role ?? "labourer") as RoleKey,
-          token ?? undefined
-        );
-      } else {
-        ensureProfile(
-          viewUserId,
-          viewRole === "labourer" ? "Labourer" : "Manager",
-          viewRole,
-          token ?? undefined
-        );
-      }
+    if (user) {
+      ensureProfile(user.id, user.username ?? "You", "client", token ?? undefined);
     }
-  }, [viewUserId, profiles, isOwn, user, ensureProfile, viewRole, token]);
+  }, [user, ensureProfile, token]);
 
-  const profile = profiles[viewUserId];
+  const profile = profiles[userId];
 
   const [editing, setEditing] = useState(false);
-  const [name, setName] = useState(
-    profile?.name ?? (isOwn ? user?.username ?? "You" : "Name")
-  );
-  const [headline, setHeadline] = useState(profile?.headline ?? "Looking for work");
-  const [location, setLocation] = useState(profile?.location ?? "Brighton, UK");
+  const [name, setName] = useState(profile?.name ?? user?.username ?? "You");
+  const [headline, setHeadline] = useState(profile?.headline ?? "Hiring now");
+  const [location, setLocation] = useState(profile?.location ?? "London, UK");
   const [company, setCompany] = useState(profile?.company ?? "");
   const [bio, setBio] = useState(profile?.bio ?? "");
 
@@ -97,21 +57,17 @@ export default function LabourerProfileDetails() {
   useEffect(() => {
     if (profile) {
       setName(profile.name);
-      setHeadline(profile.headline ?? "Looking for work");
+      setHeadline(profile.headline ?? "Hiring now");
       setLocation(profile.location ?? "");
       setCompany(profile.company ?? "");
       setBio(profile.bio ?? "");
     }
   }, [profile]);
 
-  useEffect(() => {
-    if (!isOwn) setEditing(false);
-  }, [isOwn, viewUserId]);
-
   const save = () => {
-    if (!user || !isOwn) return;
+    if (!user) return;
     updateProfile(
-      viewUserId,
+      user.id,
       {
         name: name.trim(),
         headline: headline.trim(),
@@ -132,14 +88,14 @@ export default function LabourerProfileDetails() {
     });
     if (!res.canceled && res.assets?.length) {
       const localUri = res.assets[0].uri;
-      updateProfile(viewUserId, { [field]: localUri } as any);
+      updateProfile(userId, { [field]: localUri } as any);
       if (token) {
         try {
           const url =
             field === "avatarUri"
-              ? await uploadAvatar(viewUserId, localUri, token)
-              : await uploadBanner(viewUserId, localUri, token);
-          updateProfile(viewUserId, { [field]: url } as any);
+              ? await uploadAvatar(userId, localUri, token)
+              : await uploadBanner(userId, localUri, token);
+          updateProfile(userId, { [field]: url } as any);
         } catch (err) {
           console.warn(err);
         }
@@ -154,13 +110,13 @@ export default function LabourerProfileDetails() {
       mediaTypes: ImagePicker.MediaTypeOptions.Images,
     });
     if (!res.canceled && res.assets?.length) {
-      updateQualification(viewUserId, id, { imageUri: res.assets[0].uri }, token ?? undefined);
+      updateQualification(userId, id, { imageUri: res.assets[0].uri }, token ?? undefined);
     }
   };
 
   const addQual = () => {
     addQualification(
-      viewUserId,
+      userId,
       {
         id: String(Date.now()),
         title: "New Qualification",
@@ -192,46 +148,25 @@ export default function LabourerProfileDetails() {
         keyboardVerticalOffset={Platform.OS === "ios" ? insets.top : 0}
       >
         <View style={{ flex: 1 }}>
-          {/* FIXED Top bar (outside the ScrollView) */}
-          <View style={[styles.topBar, { paddingTop: insets.top + 6 }]}>
-            <Pressable
-              onPress={() => {
-                if (from === "chat") {
-                  router.back();
-                } else if (backJobId) {
-                  const dest = from === "jobs" ? "/(labourer)/jobs" : "/(labourer)/map";
-                  router.replace({ pathname: dest, params: { jobId: String(backJobId) } });
-                } else {
-                  router.replace(BACK_TO);
-                }
-              }}
-              hitSlop={12}
-            >
-              <Ionicons name="chevron-back" size={24} color="#111" />
+          {/* Top bar */}
+          <View style={[styles.topBar, { paddingTop: insets.top + 6 }]}> 
+            <Pressable onPress={() => router.back()} hitSlop={12}> 
+              <Ionicons name="chevron-back" size={24} color="#111" /> 
             </Pressable>
 
             <Text style={styles.topTitle}>Profile</Text>
 
-            {isOwn ? (
-              <Pressable onPress={() => setEditing((e) => !e)} hitSlop={12}>
-                <Ionicons name={editing ? "close" : "pencil"} size={22} color="#6B7280" />
-              </Pressable>
-            ) : (
-              <View style={{ width: 22 }} />
-            )}
+            <Pressable onPress={() => setEditing((e) => !e)} hitSlop={12}>
+              <Ionicons name={editing ? "close" : "pencil"} size={22} color="#6B7280" />
+            </Pressable>
           </View>
 
           <ScrollView
-            contentContainerStyle={{
-              paddingBottom: Math.max(insets.bottom, 16) + 24,
-            }}
+            contentContainerStyle={{ paddingBottom: Math.max(insets.bottom, 16) + 24 }}
             keyboardShouldPersistTaps="handled"
           >
             {/* Banner */}
-            <Pressable
-              onPress={() => editing && pickImage("bannerUri")}
-              disabled={!editing}
-            >
+            <Pressable onPress={() => editing && pickImage("bannerUri")} disabled={!editing}>
               <Image
                 source={{
                   uri:
@@ -242,23 +177,16 @@ export default function LabourerProfileDetails() {
               />
             </Pressable>
 
-            {/* Avatar with silhouette fallback */}
+            {/* Avatar */}
             <View style={styles.avatarWrap}>
-              <Pressable
-                onPress={() => editing && pickImage("avatarUri")}
-                disabled={!editing}
-              >
+              <Pressable onPress={() => editing && pickImage("avatarUri")} disabled={!editing}>
                 {profile.avatarUri ? (
                   <Image source={{ uri: profile.avatarUri }} style={styles.avatar} />
                 ) : (
                   <View
                     style={[
                       styles.avatar,
-                      {
-                        alignItems: "center",
-                        justifyContent: "center",
-                        backgroundColor: "#E5E7EB",
-                      },
+                      { alignItems: "center", justifyContent: "center", backgroundColor: "#E5E7EB" },
                     ]}
                   >
                     <Ionicons name="person" size={28} color="#9CA3AF" />
@@ -271,18 +199,8 @@ export default function LabourerProfileDetails() {
             <View style={styles.card}>
               {editing ? (
                 <>
-                  <TextInput
-                    value={name}
-                    onChangeText={setName}
-                    style={styles.input}
-                    placeholder="Name"
-                  />
-                  <TextInput
-                    value={headline}
-                    onChangeText={setHeadline}
-                    style={styles.input}
-                    placeholder="Headline"
-                  />
+                  <TextInput value={name} onChangeText={setName} style={styles.input} placeholder="Name" />
+                  <TextInput value={headline} onChangeText={setHeadline} style={styles.input} placeholder="Headline" />
                 </>
               ) : (
                 <>
@@ -297,38 +215,16 @@ export default function LabourerProfileDetails() {
               )}
 
               <View style={styles.metaGrid}>
-                <Meta
-                  icon="location-outline"
-                  label="Based in"
-                  value={!editing && hasLocation ? profile.location : undefined}
-                />
-                <Meta
-                  icon="business-outline"
-                  label="Company"
-                  value={!editing && hasCompany ? profile.company : undefined}
-                />
-                <Meta
-                  icon="construct-outline"
-                  label="Jobs Completed"
-                  value={String(profile.jobsCompleted ?? 0)}
-                />
-                <Meta icon="person-outline" label="Role" value="Labourer" />
+                <Meta icon="location-outline" label="Based in" value={!editing && hasLocation ? profile.location : undefined} />
+                <Meta icon="business-outline" label="Company" value={!editing && hasCompany ? profile.company : undefined} />
+                <Meta icon="construct-outline" label="Jobs Completed" value={String(profile.jobsCompleted ?? 0)} />
+                <Meta icon="person-outline" label="Role" value="Client" />
               </View>
 
               {editing && (
                 <>
-                  <TextInput
-                    value={location}
-                    onChangeText={setLocation}
-                    style={styles.input}
-                    placeholder="Location"
-                  />
-                  <TextInput
-                    value={company}
-                    onChangeText={setCompany}
-                    style={styles.input}
-                    placeholder="Company"
-                  />
+                  <TextInput value={location} onChangeText={setLocation} style={styles.input} placeholder="Location" />
+                  <TextInput value={company} onChangeText={setCompany} style={styles.input} placeholder="Company" />
                 </>
               )}
             </View>
@@ -385,7 +281,7 @@ export default function LabourerProfileDetails() {
                         <TextInput
                           value={q.title}
                           onChangeText={(t) =>
-                            updateQualification(viewUserId, q.id, { title: t }, token ?? undefined)
+                            updateQualification(userId, q.id, { title: t }, token ?? undefined)
                           }
                           style={styles.input}
                           placeholder="Title"
@@ -396,20 +292,16 @@ export default function LabourerProfileDetails() {
                       <Text
                         style={[
                           styles.badge,
-                          q.status === "verified"
-                            ? styles.badgeVerified
-                            : styles.badgePending,
+                          q.status === "verified" ? styles.badgeVerified : styles.badgePending,
                         ]}
                       >
-                        {q.status === "verified"
-                          ? "Verified"
-                          : "Pending verification"}
+                        {q.status === "verified" ? "Verified" : "Pending verification"}
                       </Text>
                     </View>
 
                     {editing ? (
                       <Pressable
-                        onPress={() => removeQualification(viewUserId, q.id, token ?? undefined)}
+                        onPress={() => removeQualification(userId, q.id, token ?? undefined)}
                         style={[styles.btnIcon, { borderColor: "#ef4444" }]}
                         hitSlop={6}
                       >
@@ -431,10 +323,10 @@ export default function LabourerProfileDetails() {
                 onChangeNew={setNewSkill}
                 onAdd={() => {
                   if (!newSkill.trim()) return;
-                  addSkill(viewUserId, newSkill.trim(), token ?? undefined);
+                  addSkill(userId, newSkill.trim(), token ?? undefined);
                   setNewSkill("");
                 }}
-                onRemove={(s) => removeSkill(viewUserId, s, token ?? undefined)}
+                onRemove={(s) => removeSkill(userId, s, token ?? undefined)}
               />
             )}
 
@@ -448,14 +340,14 @@ export default function LabourerProfileDetails() {
                 onChangeNew={setNewInterest}
                 onAdd={() => {
                   if (!newInterest.trim()) return;
-                  addInterest(viewUserId, newInterest.trim(), token ?? undefined);
+                  addInterest(userId, newInterest.trim(), token ?? undefined);
                   setNewInterest("");
                 }}
-                onRemove={(s) => removeInterest(viewUserId, s, token ?? undefined)}
+                onRemove={(s) => removeInterest(userId, s, token ?? undefined)}
               />
             )}
 
-            {/* Save inside the scroll */}
+            {/* Save button */}
             {editing && (
               <View
                 style={{
@@ -632,3 +524,4 @@ const styles = StyleSheet.create({
   badgeVerified: { backgroundColor: "#E9F9EE", color: "#1E7F3E" },
   badgePending: { backgroundColor: "#FFF4CC", color: "#8A6A00" },
 });
+
