+++ mobile/app/(client)/_layout.tsx
@@ -36,8 +36,7 @@ export default function ClientTabs() {
       {/* hidden routes */}
       <Tabs.Screen name="index" options={{ href: null }} />
       <Tabs.Screen name="projects/[id]" options={{ href: null }} />
-      <Tabs.Screen name="profile" options={{ href: null }} />
-      <Tabs.Screen name="profileDetails" options={{ href: null }} />
+      <Tabs.Screen name="(profile)" options={{ href: null }} />
 
       {/* visible tabs */}
       <Tabs.Screen
