+++ mobile/app/(labourer)/(profile)/subscriptions.tsx
@@ -1,6 +1,6 @@
 import ProSubscription from "@src/components/ProSubscription";
 
 export default function LabourerSubscription() {
-  return <ProSubscription profileRoute="/(labourer)/profile" />;
+  return <ProSubscription profileRoute="/(labourer)/(profile)" />;
 }
