+++ mobile/app/(labourer)/jobs.tsx
@@ -101,7 +101,7 @@ export default function Jobs() {
       const { userId, jobId } = pendingProfile;
       setPendingProfile(null);
       router.push({
-        pathname: "/(labourer)/profileDetails",
+        pathname: "/(labourer)/(profile)/profileDetails",
         params: { userId: String(userId), jobId: String(jobId), from: "jobs" },
       });
     }
