+++ mobile/app/(manager)/(profile)/subscriptions.tsx
@@ -1,6 +1,6 @@
 import ProSubscription from "@src/components/ProSubscription";
 
 export default function ManagerSubscription() {
-  return <ProSubscription profileRoute="/(manager)/profile" />;
+  return <ProSubscription profileRoute="/(manager)/(profile)" />;
 }
