+++ mobile/app/(manager)/_layout.tsx
@@ -28,10 +28,7 @@ export default function ManagerTabs() {
     >
       {/* hidden routes */}
       <Tabs.Screen name="index" options={{ href: null }} />
-      <Tabs.Screen name="profile" options={{ href: null }} />
-      <Tabs.Screen name="profileDetails" options={{ href: null }} />
-      <Tabs.Screen name="subscriptions" options={{ href: null }} />
-      <Tabs.Screen name="personal-info" options={{ href: null }} />
+      <Tabs.Screen name="(profile)" options={{ href: null }} />
 
       {/* visible tabs */}
       <Tabs.Screen
